+++ src/pages/shopping-cart/ShoppingCartPage.tsx
@@ -4,7 +4,7 @@ import { getShoppingCartData } from "../../api/cart";
 import { Header } from "../../components/common";
 import { useAPIDataContext } from "../../context/APIDataProvider";
 import CartLayout from "./cart-layout/CartLayout";
-import { useOrderListContext } from "./context/OrderListProvider";
+import { useOrderListContext } from "../../context/OrderListProvider";
 
 const ShoppingCartPage = () => {
   const { data: cartListData } = useAPIDataContext({
