+++ src/pages/shopping-cart/cart-layout/LabelPriceContainer.tsx
@@ -1,7 +1,7 @@
 import styled from "@emotion/styled";
 import Flex from "../../../components/common/Flex";
 import LabelPrice from "../../../components/common/LabelPrice";
-import { useOrderListContext } from "../context/OrderListProvider";
+import { useOrderListContext } from "../../../context/OrderListProvider";
 import { formatKRWString } from "../../../utils/formatKRWString";
 import { useAPIDataContext } from "../../../context/APIDataProvider";
 import { getShoppingCartData } from "../../../api/cart";
